test-requirements.txt
coverage>=5.0.3,<5.1
flake8>=3.7.9,<4
bandit>=1.6.2
pylint>=2.4.4,<3
pylint-django>=2.0.13,<3
beautifulsoup4>=4.8.2,<5
selenium>=3.141.0
jsonschema>=3.2.0,<4
